+++ hosts/vscode-plugin/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Auto refresh files in IDE on edit/write
 - New panel with all modified files in session
+- New panel with all TODOs in session
 - Fixed placement of Model/Agent selector
 
 ## [25.11.30] - 2025-11-30